refactor : 룰렛 돌리기 경로 이름은 roulette/play 로 변경
